Docs: Update the 4.19 release notes to match the server version
The version of the 4.19 release notes on the server was updated with
signatures and a note explaining the new tarballs vs the original,
corrupted tarballs.  That data didn't make it back to the version of
the release notes in the documentation.

Likewise, the updates in the documentation didn't get pushed to the
release directory on the website.

The website now has this version of the release notes that combines the
two. This patch does the same for the documentation folder.

Signed-off-by: Martin Roth <gaumless@gmail.com>
Change-Id: Ib2563e7fa4b8d82ad4fbb3fd3880ee62a24a8aca
Reviewed-on: https://review.coreboot.org/c/coreboot/+/75639
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Eric Lai <eric_lai@quanta.corp-partner.google.com>
Reviewed-by: Felix Singer <service+coreboot-gerrit@felixsinger.de>